Added inserts, found bug, made workaround
Added inserts, which insert code snippets into lua files based on comment delimiters.
Found out that the shell utility is unavailable in files called with "loadfile", which is core to the relative file system in use in this project. No fix found as of yet.
Changed wget to not resolve path as all files given are already relative.